:sparkles: RADIO_LIST added
