--- src/utils/constants.ts
@@ -1,12 +1,0 @@
-export const INPUT_LIST = [
-  { label: 'Name', type: 'text', name:'name', placeholder: 'Stephen King' },
-  { label: 'Email Address', type: 'email', name:'email', placeholder: 'stephenking@lorem.com' },
-  { label: 'Phone Number', type: 'number', name:'number', placeholder: '+57 313 111 1111' },
-];
-
-export const STEP_LIST = [
-  { value: '1', label: 'STEP 1', description: 'YOUR INFO' },
-  { value: '2', label: 'STEP 2', description: 'SELECT PLAN' },
-  { value: '3', label: 'STEP 3', description: 'ADD-ONS' },
-  { value: '4', label: 'STEP 4', description: 'SUMMARY' },
-];
